Split files with more than one class
